requirements.txt
PyMuPDF
python-docx
pymupdf4llm
pandas
rich
pyarrow